Refactor tcp server and db core
