+++ fold_node/src/datafold_node/tests.rs
@@ -1,3 +1,4 @@
+#![allow(dead_code)]
 use super::*;
 use tempfile::tempdir;
